feat: "Add check for existing event with same slug before creating new event"
This commit adds a new line of code that checks for an existing event with the same slug before creating a new event. This is done by querying the database with the `findUnique` method on the `prisma.event` object, passing in a `where` object with the `slug` property set to the generated slug. If an event with the same slug is found, it will be assigned to the `eventWithSameSlug` variable. This check was likely added to prevent duplicate events with the same slug from being created.