feat: "Add validation for maximum attendees limit when registering for event"
The diff adds a check to ensure that the maximum number of attendees for an event has not been exceeded when registering a new attendee. It does this by querying the database to get the event and the current number of attendees for the event, and then checks if the maximum number of attendees for the event has been reached. If it has, an error is thrown.