feat: "Add Zod schema for get attendee badge route parameters"
This commit adds a Zod schema to validate the `attendeeId` parameter for the `/attendees/:attendeeId/badge` route. The schema is defined using the `z.object` and `z.coerce.number().int()` methods to specify that the `attendeeId` should be a integer number. The response schema is left empty in this commit.