feat: Prisma database dev.db and dev.db-journal files have been modified.
The previous version of dev.db is at commit `b87bbfa` and the new version is at `03f2209`. Similarly, the previous version of dev.db-journal is at `3430299` and the new version is at `8cf77a8`.

The binary files differ, indicating changes to the data or structure of the Prisma database.