feat: "Update getEventAttendees query to optionally filter by attendee name"
The code diff adds a new optional query parameter `query` to the `getEventAttendees` function. If `query` is provided, the function will filter the attendees of an event by their name, in addition to the existing filter by `eventId`. The new query parameter is used in the `where` clause of the SQL query to filter attendees whose name contains the specified query string. The take and skip options are used to paginate the results.