feat: "Add Swagger documentation and register 'registerForEvent' and 'getEvent' routes"
This commit adds Swagger documentation to the project using the `@fastify/swagger` and `@fastify/swagger-ui` modules. It also registers the `registerForEvent` and `getEvent` routes, which were previously imported but not registered. The `jsonSchemaTransform` option is used to customize the JSON schema generation. The validator and serializer compilers are also set for the project.