feat: "Add BadRequest error handling in error-handler.ts"
This commit adds a new error handling case for BadRequest errors in the error-handler.ts file. If a BadRequest error is thrown, the status code will be set to 400 and the error message will be sent in the response. If any other error is thrown, the status code will be set to 500 and the default error message will be sent.