feat: "Update error handler to handle ZodError and BadRequestException"
This commit updates the error handler function to handle two new types of errors: ZodError and BadRequestException. If the error is an instance of ZodError, the function will return a 400 status code and send a JSON object containing an error message and a list of validation errors. If the error is an instance of BadRequestException, the function will return a 400 status code and send a JSON object containing an error message. Previously, the error handler only handled errors of type Error, so this commit adds support for handling more specific types of errors.