feat: Change event title and slug in seed.ts Update dev.db and dev.db-journal with new data
This commit contains changes to the `seed.ts` file where the title and slug of an event have been updated from "Unite Summit" to "NLW Unite" and from "unite-summit" to "nlw-unite", respectively. Additionally, the `dev.db` and `dev.db-journal` files have been updated with new data as a result of running the updated seed script.